Update for livestream.md
Added explanations for scripts. I noticed the second example for useEffect had an infinite loop. This could be a helpful example as to how this tool could be used poorly. Secondly, the useRef final example did not correctly update the number of renders in the console so I added a new version of App.js to look over. At the bottom of the page I have detailed a walkthrough of the todoList in 11 steps. I did not make it overly in depth and did not include code snippets but I can add these if it will be helpful.